Moved AwardAndSmiteDialogs folder inside Dialogs
