+++ src/modules/dashboard/components/Dialogs/AwardAndSmiteDialogs/AwardDialog/AwardDialog.tsx
