+++ src/modules/dashboard/components/Dialogs/AwardAndSmiteDialogs/AwardDialog/index.ts
