+++ src/modules/dashboard/components/Dialogs/AwardAndSmiteDialogs/ManageReputationDialogForm/ManageReputationDialogForm.tsx
@@ -33,8 +33,8 @@ import { useTransformer } from '~utils/hooks';
 import { getFormattedTokenValue } from '~utils/tokens';
 import { calculatePercentageReputation } from '~utils/reputation';
 
-import { getUserRolesForDomain } from '../../../../transformers';
-import { userHasRole } from '../../../../users/checks';
+import { getUserRolesForDomain } from '~modules/transformers';
+import { userHasRole } from '~modules/users/checks';
 
 import { ManageReputationDialogFormValues } from '../types';
 import TeamDropdownItem from './TeamDropdownItem';
