+++ src/modules/dashboard/components/Dialogs/AwardAndSmiteDialogs/SmiteDialog/SmiteDialog.tsx
