+++ src/modules/dashboard/components/Dialogs/AwardAndSmiteDialogs/SmiteDialog/index.ts
