+++ src/modules/dashboard/components/Dialogs/AwardAndSmiteDialogs/index.ts
