+++ src/modules/dashboard/components/Dialogs/AwardAndSmiteDialogs/types.ts
